Fail if we try to install on an unsupported Python version.
